Refactored some of the grid view.
