--- src/components/index.ts
@@ -1,1 +1,0 @@
-export {Game} from './Game';
