+++ src/grid/Grid.css
@@ -1,11 +1,11 @@
-.Grid{
+.Grid {
     background-color: white;
     width: 100%;
     height: 100%;
-    padding: 10px;;
+    padding: 10px;
 }
 
-.Square{
+.Square {
     background: #fff;
     border: 1px solid #999;
     float: left;
@@ -20,8 +20,8 @@
     width: 34px;
 }
 
-.board-row{
+.Grid__row {
     clear: both;
     content: "";
     display: table;
-  }
+}
